tools/fetch_data/requirements_03.txt
requests
paramiko
scp
beautifulsoup4
python-dotenv